+++ src/pc/gfx/gfx_rendering_api.h
@@ -28,7 +28,7 @@ struct GfxRenderingAPI {
     void (*set_uniform)(struct ShaderProgram *prg, const char *name, ShaderUniformType type, const void *data, uint32_t numElements);
     uint32_t (*new_texture)(void);
     void (*select_texture)(int tile, uint32_t texture_id);
-    void (*bind_texture_raw)(int tile, uint32_t texture_id);
+    void (*bind_texture_raw)(int tile, uint64_t texture_id); // needs to be 64 for dx11 pointer
     void (*upload_texture)(const uint8_t *rgba32_buf, int width, int height);
     void (*set_sampler_parameters)(int sampler, bool linear_filter, uint32_t cms, uint32_t cmt);
     void (*set_depth_test)(bool depth_test);
